Add files via upload
Car Racing Game is written in HTML, CSS, and JavaScript. Talking about the gameplay, the main objective of this game is to have an endless race without failing. There are no time limits during gameplay, all you have to do is dodge the car from the white objects (other cars).

In this Car Racing game, the player has to race the car without colliding with other cars. The player has to avoid blue square-shaped color objects. The main objective of this game is to score the highest points as much as you can. Talking about the PC controls of this game, it’s all simple. All you have to use your arrow keys for movement. This is a simple game using JavaScript. All the gaming functions are from Javascript whereas HTML and CSS are for the layouts.